long

+++ bot/src/main/java/org/example/bot/dto/LinkUpdateRequest.java
@@ -2,5 +2,5 @@ package org.example.bot.dto;
 
 import java.util.List;
 
-public record LinkUpdateRequest(Integer id, String url, String descriptions, List<Integer> tgChatIds) {
+public record LinkUpdateRequest(Long id, String url, String descriptions, List<Long> tgChatIds) {
 }
