+++ scrapper/src/main/java/org/example/scrapper/dto/LinkResponse.java
@@ -1,4 +1,4 @@
 package org.example.scrapper.dto;
 
-public record LinkResponse(int id, String url) {
+public record LinkResponse(Long id, String url) {
 }
